code/exercises/reddit/finance/frodo_investment/requirements.txt
pyyaml
re
math